# server/ui/api/v1/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "OpenAPI 2.0 documentation for [MessagePit](https://github.com/coreydaley/messagepit).",
    "title": "MessagePit API",
    "contact": {
      "name": "GitHub",
      "url": "https://github.com/coreydaley/messagepit"
    },
    "license": {
      "name": "MIT license",
      "url": "https://github.com/coreydaley/messagepit/blob/main/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/message/{ID}/part/{PartID}/thumb": {
      "get": {
        "description": "This will return a cropped 180x120 JPEG thumbnail of an image attachment.\nIf the image is smaller than 180x120 then the image is padded. If the attachment is not an image then a blank image is returned.\n\nThe ID can be set to `latest` to return the latest message.",
        "produces": [
          "image/jpeg"
        ],
        "schemes": [
          "http",
          "https"
        ],
        "tags": [
          "message"
        ],
        "summary": "Get an attachment image thumbnail",
        "operationId": "ThumbnailParams",
        "parameters": [
          {
            "type": "string",
            "description": "Message database ID or \"latest\"",
            "name": "ID",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Attachment part ID",
            "name": "PartID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/BinaryResponse"
          },
          "400": {
            "$ref": "#/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
...
  "responses": {
...
    "BinaryResponse": {
      "description": "Binary data response which inherits the attachment's content type.",
      "schema": {
        "type": "string"
      }
    },
...
    "ErrorResponse": {
      "description": "Server error will return with a 400 status code\nwith the error message in the body",
      "schema": {
        "type": "string"
      }
    },
...
  }
}